Fixed bug concerning evaluate_metrics fn when Y matrix with both catgs and cont factors is manipulated before function call

+++ src/pyoptex/doe/cost_optimal/evaluate.py
@@ -41,7 +41,7 @@ def evaluate_metrics(Y, params, metrics):
 
     # Transform Y to numpy
     col_names = [str(f.name) for f in params.factors]
-    Y = Y[col_names].to_numpy()
+    Y = Y[col_names].astype(float).to_numpy()
 
     # Encode the design
     Y = encode_design(Y, params.effect_types, params.coords)
